# apps/backend/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/iam/roles/{id}/permissions/{permissionId}": {
      "delete": {
        "operationId": "RoleController_removePermission",
        "summary": "Remove a permission from a role",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "permissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Permission removed from role"
          }
        },
        "tags": [
          "iam"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/iam/permissions": {
      "get": {
        "operationId": "PermissionController_findAll",
        "summary": "List all permissions (global)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "iam"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "PermissionController_create",
        "summary": "Create a new permission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permission created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          }
        },
        "tags": [
          "iam"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/iam/permissions/{id}": {
      "get": {
        "operationId": "PermissionController_findOne",
        "summary": "Get permission by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          }
        },
        "tags": [
          "iam"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "PermissionController_update",
        "summary": "Update a permission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          }
        },
        "tags": [
          "iam"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "PermissionController_delete",
        "summary": "Delete a permission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Permission deleted"
          }
        },
        "tags": [
          "iam"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/iam/permissions/by-resource/{resource}": {
      "get": {
        "operationId": "PermissionController_findByResource",
        "summary": "Get permissions by resource",
        "description": "Returns all permissions for a specific resource",
        "parameters": [
          {
            "name": "resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "iam"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/iam/permissions/by-action/{action}": {
      "get": {
        "operationId": "PermissionController_findByAction",
        "summary": "Get permissions by action",
        "description": "Returns all permissions for a specific action",
        "parameters": [
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "iam"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/iam/permissions/{id}/usage": {
      "get": {
        "operationId": "PermissionController_getPermissionUsage",
        "summary": "Get permission usage",
        "description": "Returns all roles that have this permission assigned",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of roles with this permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "iam"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TENX API",
    "description": "Multi-tenant SaaS ERP backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
      "Role": {
        "type": "object",
        "properties": {}
      },
...
      "Permission": {
        "type": "object",
        "properties": {}
      },
...
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "resource": {
            "type": "string",
            "maxLength": 100,
            "example": "user"
          },
          "action": {
            "type": "string",
            "maxLength": 50,
            "example": "read"
          },
          "description": {
            "type": "string",
            "example": "Permission to read user data"
          },
          "category": {
            "type": "string",
            "maxLength": 50,
            "example": "Users"
          }
        },
        "required": [
          "resource",
          "action"
        ]
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "resource": {
            "type": "string",
            "maxLength": 100
          },
          "action": {
            "type": "string",
            "maxLength": 50
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "maxLength": 50
          }
        }
      },
...
    }
  }
}